Factor Tags::Variables from Variables.hpp
