+++ src/Elliptic/Systems/Elasticity/FirstOrderSystem.hpp
@@ -6,6 +6,7 @@
 #include <cstddef>
 
 #include "DataStructures/Tensor/EagerMath/Magnitude.hpp"
+#include "DataStructures/VariablesTag.hpp"
 #include "Elliptic/Systems/Elasticity/Equations.hpp"
 #include "Elliptic/Systems/Elasticity/Tags.hpp"
 #include "NumericalAlgorithms/LinearOperators/PartialDerivatives.hpp"
@@ -14,11 +15,6 @@
 #include "Utilities/TMPL.hpp"
 
 /// \cond
-namespace Tags {
-template <typename>
-class Variables;
-}  // namespace Tags
-
 namespace LinearSolver {
 namespace Tags {
 template <typename>
