+++ src/Elliptic/Systems/Poisson/FirstOrderSystem.hpp
@@ -9,6 +9,7 @@
 #include <cstddef>
 
 #include "DataStructures/Tensor/EagerMath/Magnitude.hpp"
+#include "DataStructures/VariablesTag.hpp"
 #include "Elliptic/Systems/Poisson/Equations.hpp"
 #include "Elliptic/Systems/Poisson/Geometry.hpp"
 #include "Elliptic/Systems/Poisson/Tags.hpp"
@@ -18,11 +19,6 @@
 #include "Utilities/TMPL.hpp"
 
 /// \cond
-namespace Tags {
-template <typename>
-class Variables;
-}  // namespace Tags
-
 namespace LinearSolver {
 namespace Tags {
 template <typename>
