+++ src/Evolution/Systems/Cce/Components/CharacteristicEvolution.hpp
@@ -5,6 +5,7 @@
 
 #include <type_traits>
 
+#include "DataStructures/VariablesTag.hpp"
 #include "Evolution/Systems/Cce/Actions/BoundaryComputeAndSendToEvolution.hpp"
 #include "Evolution/Systems/Cce/Actions/CalculateScriInputs.hpp"
 #include "Evolution/Systems/Cce/Actions/CharacteristicEvolutionBondiCalculations.hpp"
