+++ src/NumericalAlgorithms/Interpolation/InitializeInterpolationTarget.hpp
@@ -9,6 +9,7 @@
 
 #include "DataStructures/DataBox/DataBox.hpp"
 #include "DataStructures/DataBox/DataBoxTag.hpp"
+#include "DataStructures/VariablesTag.hpp"
 #include "NumericalAlgorithms/Interpolation/Tags.hpp"  // IWYU pragma: keep
 #include "Utilities/Requires.hpp"
 #include "Utilities/TMPL.hpp"
@@ -21,10 +22,6 @@ namespace Parallel {
 template <typename Metavariables>
 class ConstGlobalCache;
 }  // namespace Parallel
-namespace Tags {
-template <typename TagsList>
-struct Variables;
-}  // namespace Tags
 /// \endcond
 
 namespace intrp {
